pyproject.toml
[project]
name = "flower-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "tensorflow==2.17.0",
    "numpy==1.26.4",
    "pandas==2.2.2",
    "pillow==10.4.0",
    "openpyxl==3.1.5",
]
